test before delete service.old

+++ routines/root_user/service_definitions.sh
@@ -4,7 +4,10 @@ function setup_service_definitions {
 cd /opt/engines/etc
 if test -d services
 	then
-	 rm -r services.old
+		if test -d services.old
+		 then
+	 		rm -r services.old
+	 	fi
 		mv services services.old
 	fi
 git clone https://github.com/EnginesOS/ServiceDefinitions services
